packages/healops_opentelemetry_python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "healops_opentelemetry"
version = "0.1.0"
description = "HealOps OpenTelemetry SDK for Python"
readme = "README.md"
authors = [{ name = "HealOps", email = "support@healops.ai" }]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation",
    "requests"
]
requires-python = ">=3.7"

[project.urls]
"Homepage" = "https://healops.ai"
"Bug Tracker" = "https://github.com/healops/healops-otel-python/issues"
